Code refactor for beacons, moved to a conext
