+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_RootScreen/components/Announcements.tsx
@@ -13,9 +13,9 @@ import {useAnalytics} from '@atb/analytics';
 import {AnnouncementType} from '@atb/announcements/types';
 import {isWithinTimeRange} from '@atb/utils/is-within-time-range';
 import {useNow} from '@atb/utils/use-now';
-import {useBeacons} from '@atb/beacons/use-beacons';
 import {useHasSeenShareTravelHabitsScreen} from '@atb/beacons/use-has-seen-share-travel-habits-screen';
 import Bugsnag from '@bugsnag/react-native';
+import {useBeaconsState} from '@atb/beacons/BeaconsContext';
 
 type Props = {
   style?: StyleProp<ViewStyle>;
@@ -28,7 +28,7 @@ export const Announcements = ({style: containerStyle}: Props) => {
   const style = useStyle();
   const analytics = useAnalytics();
   const now = useNow(10000);
-  const {kettleInfo} = useBeacons();
+  const {kettleInfo} = useBeaconsState();
   const [hasSeenShareTravelHabitsScreen, _] =
     useHasSeenShareTravelHabitsScreen();
